+++ internal/runtime/runtime.go
@@ -20,7 +20,49 @@ func CanUseES6(unsupportedFeatures compat.JSFeature) bool {
 }
 
 func code(isES6 bool) string {
-	// Note: The "__rest" function has a for-of loop which requires ES6, but
+	// Note: These helper functions used to be named similar things to the helper
+	// functions from the TypeScript compiler. However, people sometimes use these
+	// two projects in combination and TypeScript's implementation of these helpers
+	// causes name collisions. Some examples:
+	//
+	// * The "tslib" library will overwrite esbuild's helper functions if the bundled
+	//   code is run in the global scope: https://github.com/evanw/esbuild/issues/1102
+	//
+	// * Running the TypeScript compiler on esbuild's output to convert ES6 to ES5
+	//   will also overwrite esbuild's helper functions because TypeScript doesn't
+	//   change the names of its helper functions to avoid name collisions:
+	//   https://github.com/microsoft/TypeScript/issues/43296
+	//
+	// These can both be considered bugs in TypeScript. However, they are unlikely
+	// to be fixed and it's simplest to just avoid using the same names to avoid
+	// these bugs. Forbidden names (from "tslib"):
+	//
+	//   __assign
+	//   __asyncDelegator
+	//   __asyncGenerator
+	//   __asyncValues
+	//   __await
+	//   __awaiter
+	//   __classPrivateFieldGet
+	//   __classPrivateFieldSet
+	//   __createBinding
+	//   __decorate
+	//   __exportStar
+	//   __extends
+	//   __generator
+	//   __importDefault
+	//   __importStar
+	//   __makeTemplateObject
+	//   __metadata
+	//   __param
+	//   __read
+	//   __rest
+	//   __spread
+	//   __spreadArray
+	//   __spreadArrays
+	//   __values
+	//
+	// Note: The "__objRest" function has a for-of loop which requires ES6, but
 	// transforming destructuring to ES5 isn't even supported so it's ok.
 	text := `
 		var __create = Object.create
@@ -38,7 +80,7 @@ func code(isES6 bool) string {
 			? __defProp(obj, key, {enumerable: true, configurable: true, writable: true, value})
 			: obj[key] = value
 
-		export var __assign = (a, b) => {
+		export var __objSpread = (a, b) => {
 			for (var prop in b ||= {})
 				if (__hasOwnProp.call(b, prop))
 					__defNormalProp(a, prop, b[prop])
@@ -72,7 +114,7 @@ func code(isES6 bool) string {
 
 		// For object rest patterns
 		export var __restKey = key => typeof key === 'symbol' ? key : key + ''
-		export var __rest = (source, exclude) => {
+		export var __objRest = (source, exclude) => {
 			var target = {}
 			for (var prop in source)
 				if (__hasOwnProp.call(source, prop) && exclude.indexOf(prop) < 0)
@@ -110,7 +152,7 @@ func code(isES6 bool) string {
 			for (var name in all)
 				__defProp(target, name, { get: all[name], enumerable: true })
 		}
-		export var __exportStar = (target, module, desc) => {
+		export var __reExport = (target, module, desc) => {
 			if (module && typeof module === 'object' || typeof module === 'function')
 	`
 
@@ -137,7 +179,7 @@ func code(isES6 bool) string {
 
 		// Converts the module from CommonJS to ES6 if necessary
 		export var __toModule = module => {
-			return __exportStar(__markAsModule(
+			return __reExport(__markAsModule(
 				__defProp(
 					module != null ? __create(__getProtoOf(module)) : {},
 					'default',
@@ -157,7 +199,7 @@ func code(isES6 bool) string {
 		// - kind === undefined: class
 		// - kind === 1: method, parameter
 		// - kind === 2: field
-		export var __decorate = (decorators, target, key, kind) => {
+		export var __decorateClass = (decorators, target, key, kind) => {
 			var result = kind > 1 ? void 0 : kind ? __getOwnPropDesc(target, key) : target
 			for (var i = decorators.length - 1, decorator; i >= 0; i--)
 				if (decorator = decorators[i])
@@ -166,7 +208,7 @@ func code(isES6 bool) string {
 				__defProp(target, key, result)
 			return result
 		}
-		export var __param = (index, decorator) => (target, key) => decorator(target, key, index)
+		export var __decorateParam = (index, decorator) => (target, key) => decorator(target, key, index)
 
 		// For class members
 		export var __publicField = (obj, key, value) => {
@@ -210,9 +252,7 @@ func code(isES6 bool) string {
 						reject(e)
 					}
 				}
-				var step = result => {
-					return result.done ? resolve(result.value) : Promise.resolve(result.value).then(fulfilled, rejected)
-				}
+				var step = x => x.done ? resolve(x.value) : Promise.resolve(x.value).then(fulfilled, rejected)
 				step((generator = generator.apply(__this, __arguments)).next())
 			})
 		}
@@ -259,12 +299,12 @@ var ES5Source = logger.Source{
 // The TypeScript decorator transform behaves similar to the official
 // TypeScript compiler.
 //
-// One difference is that the "__decorate" function doesn't contain a reference
+// One difference is that the "__decorateClass" function doesn't contain a reference
 // to the non-existent "Reflect.decorate" function. This function was never
 // standardized and checking for it is wasted code (as well as a potentially
 // dangerous cause of unintentional behavior changes in the future).
 //
-// Another difference is that the "__decorate" function doesn't take in an
+// Another difference is that the "__decorateClass" function doesn't take in an
 // optional property descriptor like it does in the official TypeScript
 // compiler's support code. This appears to be a dead code path in the official
 // support code that is only there for legacy reasons.
@@ -276,7 +316,7 @@ var ES5Source = logger.Source{
 //   // TypeScript                      // JavaScript
 //   @dec                               let C = class {
 //   class C {                          };
-//   }                                  C = __decorate([
+//   }                                  C = __decorateClass([
 //                                        dec
 //                                      ], C);
 //
@@ -286,7 +326,7 @@ var ES5Source = logger.Source{
 //   class C {                          class C {
 //     @dec                               foo() {}
 //     foo() {}                         }
-//   }                                  __decorate([
+//   }                                  __decorateClass([
 //                                        dec
 //                                      ], C.prototype, 'foo', 1);
 //
@@ -296,8 +336,8 @@ var ES5Source = logger.Source{
 //   class C {                          class C {
 //     foo(@dec bar) {}                   foo(bar) {}
 //   }                                  }
-//                                      __decorate([
-//                                        __param(0, dec)
+//                                      __decorateClass([
+//                                        __decorateParam(0, dec)
 //                                      ], C.prototype, 'foo', 1);
 //
 // ============================= Field decorator ==============================
@@ -308,6 +348,6 @@ var ES5Source = logger.Source{
 //     foo = 123                            this.foo = 123
 //   }                                    }
 //                                      }
-//                                      __decorate([
+//                                      __decorateClass([
 //                                        dec
 //                                      ], C.prototype, 'foo', 2);